requirements.txt
python-dotenv
openai
chainlit
azure-core
azure-ai-documentintelligence
requests
numpy